remove some depence
